+++ gaios/DrawerNetworkSelection/Controllers/DrawerNetworkSelectionViewController.swift
@@ -25,12 +25,6 @@ class DrawerNetworkSelectionViewController: UIViewController {
     var footerH: CGFloat = 54.0
     var isAnimating = false
 
-    private var ephAccounts: [Account] {
-        AccountsRepository.shared.ephAccounts.filter { account in
-            account.isEphemeral && !WalletsRepository.shared.wallets.filter {$0.key == account.id }.isEmpty
-        }
-    }
-
     override func viewDidLoad() {
         super.viewDidLoad()
 
@@ -64,7 +58,7 @@ class DrawerNetworkSelectionViewController: UIViewController {
         case .swWallet:
             return AccountsRepository.shared.swAccounts[indexPath.row]
         case .ephWallet:
-            return ephAccounts[indexPath.row]
+            return AccountsRepository.shared.ephAccounts[indexPath.row]
         case .hwWallet:
             return AccountsRepository.shared.hwVisibleAccounts[indexPath.row]
         default:
@@ -114,7 +108,7 @@ extension DrawerNetworkSelectionViewController: UITableViewDataSource, UITableVi
         case .swWallet:
             return AccountsRepository.shared.swAccounts.count
         case .ephWallet:
-            return ephAccounts.count
+            return AccountsRepository.shared.ephAccounts.count
         case .hwWallet:
             return AccountsRepository.shared.hwVisibleAccounts.count
         default:
@@ -135,7 +129,7 @@ extension DrawerNetworkSelectionViewController: UITableViewDataSource, UITableVi
                 return cell
             }
         case .ephWallet:
-            let account = ephAccounts[indexPath.row]
+            let account = AccountsRepository.shared.ephAccounts[indexPath.row]
             if let cell = tableView.dequeueReusableCell(withIdentifier: "WalletListCell") as? WalletListCell {
                 cell.configure(item: account,
                                indexPath: indexPath,