+++ gaios/HWFlow/Managers/BleHwManager.swift
@@ -207,6 +207,10 @@ class BleHwManager {
                    fullRestore: false,
                    parentWalletId: walletId
                 )
+                // block to login on bip32 passphrase jade wallet
+                if let hashId1 = account.walletHashId, let hashId2 = walletManager.account.walletHashId, hashId1 != hashId2 {
+                    throw GaError.GenericError("Wallet hash is different from the previous wallet")
+                }
             }
         } catch {
             let text = toBleError(error, network: nil).localizedDescription