+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 - Sign messages with Jade
+- Improve ephemeral wallet with a bip39 passphrase login
 
 ## [5.0.9] - 2025-09-08